function_app/requirements.txt
azure-functions==1.21.3
azure-identity==1.19.0
azure-mgmt-resource==23.2.0
azure-mgmt-authorization==4.0.0
